+++ src/interfaces/ISablierV2Linear.sol
@@ -18,7 +18,7 @@ interface ISablierV2Linear is ISablierV2 {
     /// @notice Reads the cliff time of the stream.
     /// @param streamId The id of the stream to make the query for.
     /// @return cliffTime The cliff time of the stream.
-    function getCliffTime(uint256 streamId) external view returns (uint64 cliffTime);
+    function getCliffTime(uint256 streamId) external view returns (uint40 cliffTime);
 
     /// @notice Reads the stream struct.
     /// @param streamId The id of the stream to make the query for.
@@ -55,11 +55,11 @@ interface ISablierV2Linear is ISablierV2 {
     function create(
         address sender,
         address recipient,
-        uint256 depositAmount,
+        uint128 depositAmount,
         address token,
-        uint64 startTime,
-        uint64 cliffTime,
-        uint64 stopTime,
+        uint40 startTime,
+        uint40 cliffTime,
+        uint40 stopTime,
         bool cancelable
     ) external returns (uint256 streamId);
 
@@ -83,10 +83,10 @@ interface ISablierV2Linear is ISablierV2 {
     function createWithDuration(
         address sender,
         address recipient,
-        uint256 depositAmount,
+        uint128 depositAmount,
         address token,
-        uint64 cliffDuration,
-        uint64 totalDuration,
+        uint40 cliffDuration,
+        uint40 totalDuration,
         bool cancelable
     ) external returns (uint256 streamId);
 }
