+++ src/libraries/Events.sol
@@ -16,7 +16,7 @@ library Events {
     /// @param recipient The address of the recipient.
     /// @param withdrawAmount The amount of tokens withdrawn to the recipient, in units of the token's decimals.
     /// @param returnAmount The amount of tokens returned to the sender, in units of the token's decimals.
-    event Cancel(uint256 indexed streamId, address indexed recipient, uint256 withdrawAmount, uint256 returnAmount);
+    event Cancel(uint256 indexed streamId, address indexed recipient, uint128 withdrawAmount, uint128 returnAmount);
 
     /// @notice Emitted when a linear stream is created.
     /// @param streamId The id of the newly created stream.
@@ -34,11 +34,11 @@ library Events {
         address indexed funder,
         address indexed sender,
         address indexed recipient,
-        uint256 depositAmount,
+        uint128 depositAmount,
         address token,
-        uint64 startTime,
-        uint64 cliffTime,
-        uint64 stopTime,
+        uint40 startTime,
+        uint40 cliffTime,
+        uint40 stopTime,
         bool cancelable
     );
 
@@ -60,13 +60,13 @@ library Events {
         address indexed funder,
         address indexed sender,
         address indexed recipient,
-        uint256 depositAmount,
+        uint128 depositAmount,
         address token,
-        uint64 startTime,
-        uint64 stopTime,
-        uint256[] segmentAmounts,
+        uint40 startTime,
+        uint40 stopTime,
+        uint128[] segmentAmounts,
         SD59x18[] segmentExponents,
-        uint64[] segmentMilestones,
+        uint40[] segmentMilestones,
         bool cancelable
     );
 
@@ -78,5 +78,5 @@ library Events {
     /// @param streamId The id of the stream.
     /// @param recipient The address of the recipient.
     /// @param amount The amount of tokens withdrawn, in units of the token's decimals.
-    event Withdraw(uint256 indexed streamId, address indexed recipient, uint256 amount);
+    event Withdraw(uint256 indexed streamId, address indexed recipient, uint128 amount);
 }
